Support autodetect file xheader
make header as a string parameter while still support boolean as back
compaitable. The default value of 'header' option sets to 'auto',
which will check the first field of the first row to determine if
there is a header. If the field contains a number, we believe there
is not header but data appears in first row, otherwise there is a header.